+++ src/dist_s1/data_models/output_models.py
@@ -7,39 +7,10 @@ import numpy as np
 import rasterio
 from pydantic import BaseModel, field_validator, model_validator
 
-from dist_s1.constants import PRODUCT_VERSION
+from dist_s1.constants import EXPECTED_FORMAT_STRING, PRODUCT_VERSION, TIF_LAYERS, TIF_LAYER_DTYPES
 from dist_s1.rio_tools import get_mgrs_profile
 from dist_s1.water_mask import apply_water_mask
 
-
-TIF_LAYER_DTYPES = {
-    'GEN-DIST-STATUS': 'uint8',
-    'GEN-METRIC': 'float32',
-    'GEN-DIST-STATUS-ACQ': 'uint8',
-    'GEN-METRIC-MAX': 'float32',
-    'GEN-DIST-CONF': 'int16',
-    'GEN-DIST-DATE': 'int16',
-    'GEN-DIST-COUNT': 'uint8',
-    'GEN-DIST-PERC': 'uint8',
-    'GEN-DIST-DUR': 'int16',
-    'GEN-DIST-LAST-DATE': 'int16',
-}
-COMP_BASELINE_LAYERS = {'GEN-DIST-STATUS', 'GEN-METRIC', 'GEN-DIST-STATUS-ACQ'}
-CONF_DB_LAYERS = {
-    'GEN-DIST-STATUS',
-    'GEN-METRIC',
-    'GEN-METRIC-MAX',
-    'GEN-DIST-CONF',
-    'GEN-DIST-DATE',
-    'GEN-DIST-COUNT',
-    'GEN-DIST-PERC',
-    'GEN-DIST-DUR',
-    'GEN-DIST-LAST-DATE',
-}
-TIF_LAYERS = TIF_LAYER_DTYPES.keys()
-EXPECTED_FORMAT_STRING = (
-    'OPERA_L3_DIST-ALERT-S1_T{mgrs_tile_id}_{acq_datetime}_{proc_datetime}_S1_30_v{PRODUCT_VERSION}'
-)
 
 PRODUCT_TAGS_FOR_EQUALITY = [
     'pre_rtc_opera_ids',
@@ -178,7 +149,7 @@ class ProductFileData(BaseModel):
         return True, 'Files match perfectly.'
 
 
-class ProductDirectoryData(BaseModel):
+class DistS1ProductDirectory(BaseModel):
     product_name: str
     dst_dir: Path | str
     tif_layer_dtypes: ClassVar[dict[str, str]] = dict(TIF_LAYER_DTYPES)
@@ -213,10 +184,14 @@ class ProductDirectoryData(BaseModel):
         layer_dict['browse'] = self.product_dir_path / f'{self.product_name}.png'
         return layer_dict
 
+    @property
+    def acq_datetime(self) -> datetime:
+        return self.product_name.acq_date_time
+
     def validate_layer_paths(self) -> bool:
         failed_layers = []
         for layer, path in self.layer_path_dict.items():
-            if layer not in COMP_BASELINE_LAYERS:
+            if layer not in TIF_LAYERS:
                 continue
             if not path.exists():
                 warn(f'Layer {layer} does not exist at path: {path}', UserWarning)
@@ -226,33 +201,7 @@ class ProductDirectoryData(BaseModel):
     def validate_tif_layer_dtypes(self) -> bool:
         failed_layers = []
         for layer, path in self.layer_path_dict.items():
-            if layer not in COMP_BASELINE_LAYERS:
-                continue
-            if path.suffix == '.tif':
-                with rasterio.open(path) as src:
-                    if src.dtypes[0] != TIF_LAYER_DTYPES[layer]:
-                        warn(
-                            f'Layer {layer} has incorrect dtype: {src.dtypes[0]}; should be: {TIF_LAYER_DTYPES[layer]}',
-                            UserWarning,
-                        )
-                        failed_layers.append(layer)
-        return len(failed_layers) == 0
-
-    # Validate CONF DB layers. To do: fix the repeated code for a better method
-    def validate_conf_db_layer_paths(self) -> bool:
-        failed_layers = []
-        for layer, path in self.layer_path_dict.items():
-            if layer not in CONF_DB_LAYERS:
-                continue
-            if not path.exists():
-                warn(f'Layer {layer} does not exist at path: {path}', UserWarning)
-                failed_layers.append(layer)
-        return len(failed_layers) == 0
-
-    def validate_conf_db_tif_layer_dtypes(self) -> bool:
-        failed_layers = []
-        for layer, path in self.layer_path_dict.items():
-            if layer not in CONF_DB_LAYERS:
+            if layer not in TIF_LAYERS:
                 continue
             if path.suffix == '.tif':
                 with rasterio.open(path) as src:
@@ -265,7 +214,7 @@ class ProductDirectoryData(BaseModel):
         return len(failed_layers) == 0
 
     def __eq__(
-        self, other: 'ProductDirectoryData', *, rtol: float = 1e-05, atol: float = 1e-05, equal_nan: bool = True
+        self, other: 'DistS1ProductDirectory', *, rtol: float = 1e-05, atol: float = 1e-05, equal_nan: bool = True
     ) -> bool:
         """Compare two ProductDirectoryData instances for equality.
 
@@ -352,7 +301,7 @@ class ProductDirectoryData(BaseModel):
         return equality
 
     @classmethod
-    def from_product_path(cls, product_dir_path: Path | str) -> 'ProductDirectoryData':
+    def from_product_path(cls, product_dir_path: Path | str) -> 'DistS1ProductDirectory':
         """Create a ProductDirectoryData instance from an existing product directory path.
 
         Parameters
@@ -396,7 +345,7 @@ class ProductDirectoryData(BaseModel):
         dst_dir: Path | str,
         water_mask_path: Path | str | None = None,
         overwrite: bool = False,
-    ) -> 'ProductDirectoryData':
+    ) -> 'DistS1ProductDirectory':
         """Generate a product directory with placeholder GeoTIFF files containing zeros.
 
         Parameters