+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `n_lookbacks` (which should have been called `n_baselines`) for computing confirmation within the SAS
    - No longer is there complicated accounting to keeping track of and confirming changes through baseline
 - Constant for `BASE_DATE`.
+- Extra validation for confirmation packaging
 
 ## Changed
 - Multiwindow strategy is now the default in both python API and CLI
@@ -25,6 +26,11 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   - `AlgoConfigData` is an attribute of the `RunConfigData`
   - Retrieval (and serialization) of the Algorithm Parameters can still be obtained via `get_public_attributes`.
 - Put data/path utils into `data_utils.py`.
+- Variables for confirmation processing (use snake case wherever possible)
+- Now have an output product without confirmation and a product with confirmation (for provenance)
+- De-couple confirmation and packaging in workflow logic
+- Confidence is now float32 (not int16) - this needs to happen because our metrics is a float32.
+- Updated golden dataset to reflect new changes.
 
 ## Added
 - Constants for CLI and `RunConfigData` - allows for consistent data parsing.
@@ -38,6 +44,11 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Validation for external model usage
 - Description of Runconfig Variables
 - Cloud-optimized Geotiff Format in Packaging
+- Confirmation is now contained inside its own function `confirmation.py`
+  - Exposed parameter `consecutive_nodist`, if `True` the `nocount` condition is applied (doesn't allow 2 consecutive misses).
+  - Exposed parameter `percent_reset_thresh`, it will apply reset if `percent` below threshold. 
+  - Exposed parameter `nocount_reset_thresh`, it will apply reset if `prevnocount` is above threshold.   
+- Confirmation CLI
   - Ensuring additional profile keys for in default GTiff are not present.
 
 ## Fixed